feat: SHT30 센서 추가 및 MQTT 제어 기능 구현
- SHT30 온도/습도 센서 추가 (GPIO 32, 35)
  - 온도: T[] = -66.875 + 218.75 * (VT/VDD)
  - 습도: RH = -19.7/0.54 + (100/0.54) * (VRH/VDD)
  - analogReadMilliVolts() 사용으로 ADC 보정 적용

- ADC 설정 최적화
  - 11dB 감쇠 설정 (0-3.3V 전체 범위)
  - 12비트 해상도 (0-4095)
  - NTC 필터: analogReadMilliVolts() 사용으로 변경

- MQTT 기능 강화
  - 제어 명령 파싱 (PWR, TMP, JSP, RES)
  - 온도/시간 원격 설정 + Flash 저장
  - 설정 변경 시 비프음 추가
  - publishData() 매개변수 제거 (gCUR 직접 사용)

- MQTT 데이터 전송 개선
  - 제어온도: NTC 센서 값 (measure_ntc_temp)
  - T1온도: SHT30 온도 (sht30_temp)
  - 습도: SHT30 습도 (sht30_humidity)
  - 팬전류: 실시간 측정값 (fan_current)
  - 릴레이상태: RY1~RY8 비트맵 (relay_state.u8)

- 릴레이 상태 추적 시스템
  - RELAY_DATA 구조체 추가 (RY1-RY8)
  - FAN(RY3), HEATER(RY2), DAMPER(RY4) 상태 실시간 기록
  - 과열/냉각 모드에서도 상태 동기화

- 센서 출력 통합
  - 1초마다 통합 출력: NTC, SHT30 온도, SHT30 습도
  - 개별 센서 printf 제거로 로그 간소화

- typedef.h 구조체 정리
  - measure_sht_temp/humidity  sht30_temp/humidity
  - fan_current 필드 추가
  - RELAY_DATA union 추가

+++ src/TM1638Display/TM1638Display.h
@@ -111,7 +111,7 @@ public:
   void sendToDisplay();// 전체 디스플레이 업데이트
   void beep();  // 부저 소리 (50ms)
   void processPowerButton();  // 전원 버튼 처리 (디바운싱 없음)
-  int  avr_NTC1=0;
+  float  avr_NTC1=0;
 
 private:
   uint8_t _pinSTB;
